requirements.txt
pytest==7.4.3
python-dotenv==1.0.0 